Centered rotation operation created and local minima demo found.
I created a differential operation object to complete a translation, linear rotation and a translation. This is more efficient as it avoids using the identity derivatives of translation.

I also found one of my demos of the find angles function was failing. I will investigate why.

I suspect it could be down to incorrect calculus from the change. a local minima not what tries are remove.
The latter is more likely as I have now changed the base to use an angled servo and have range up to pi