Added correction mechanism for base angle to avoid local minima
I noticed that there seemed to be less issues with local minima if the base was free to spin a full 360. So I created a dedicated object to represent a 0 to 180 base arm. It internally acts like a 0 to 360 arm but there is an algorithm so convert angles in this range to the equivalent in 0 to 180